requirements.txt
numpy
pandas
matplotlib
seaborn
pillow
scikit-learn
opencv-python
mediapipe
insightface
onnxruntime-gpu
lightning
wandb
albumentations